requirements.txt
streamlit
openai
agno
python-dotenv
groq
lance-db
yfinance
duckduckgo-search
pypdf
sqlalchemy
psycopg
pgvector
google-genai
